Implement recording of nodes, gather more data.
Former-commit-id: 18649a676f3479d71e6ed93b4d99970be67e85d2
Former-commit-id: 873a3ace76de87b2fc8fd88a5e13b9aa7fd66dd1 [formerly 8822d4411ba561d87a16f3d6270e70f9d4f43069] [formerly f2a01dde5031841d67b403c8555f133816ba2ca2 [formerly 4f9c29bf79f458d5bbf8124ef6d3ca4b32fb260d]] [formerly da54ab7ae8dd2b1151416ba8584c72805f8b3138 [formerly ae94bf660267bb58a5e7a19228c5b38724166e2f] [formerly f8f34f22c0b39a4f381ce9b0d5ada56d9948e57b [formerly c981f80e49dd2f1afccdd9a3fbfd0db79d1f1212]]]
Former-commit-id: ac5d27de31c993a3387dd6fe08d1ca0890393a8a [formerly 75eb3d6173df54d4c18e63f9eb385f57c7391d00] [formerly 6b4958f2f4a5d6353bb3d566f419b7fb93cde827 [formerly 884f594970a7b1db6c182379eba593a8a53b9232]]
Former-commit-id: 32f474822124aef9e265baf9b6dc146a613bb5af [formerly fb37013f9ea96c2ab46c0bdd0380358eb7de4801]
Former-commit-id: 9b3fca4f6c4e679baaf78075165d7429eccefa43
Former-commit-id: dc1a860e4b9824f10ef08948303174286160223a [formerly 28eab63cdbdc4344ba529cfd167c7513ff73e94d] [formerly f46d2159c31756e96e483128ca738f4886928c20 [formerly 378f1a0b09a15b2906d767bfd96bee955c010fa2]] [formerly 067db00ac3e6b1b83f28710524774cc8b8103803 [formerly 84061b5be2893cd4f06cb3ce0e1d041183fdd1f6] [formerly 378f1a0b09a15b2906d767bfd96bee955c010fa2 [formerly 6b41b6afba0e72b3a11bb423fd6c1f5c45d8247b]]]
Former-commit-id: b4d0be704c9213f5f8ef5d358e1682bc80a4ecf2 [formerly db9c624215581fa11ea4256fea2ecf42a6a529a9] [formerly 616f5b42cfbda0b4cd4f30daa01f8d4abf789375 [formerly 381deca2a306c2976888e3a6f9cab8340d1e89b3]]
Former-commit-id: 616f5b42cfbda0b4cd4f30daa01f8d4abf789375
Former-commit-id: 905dcdd2e844778aac709951a19a7121d6f4d716

+++ components/spackmanager/spack_manager.py
@@ -11,14 +11,16 @@ import shutil
 @dataclass
 class SpackManager:
     
-    handler_id  : str
-    spack_env   : dict
-    env_name    : str
-    target      : list
-    language    : list
-    packages    : dict
-    additional  : str
-    env_location: Path
+    handler_id       : str
+    spack_env        : dict
+    env_name         : str
+    target           : list
+    language         : list
+    packages         : dict
+    additional       : str
+    env_location     : Path
+    package_locations: dict
+    initialized      : bool
     
     
     def __init__(self,
@@ -26,18 +28,23 @@ class SpackManager:
                  env_name       : str,
                  spack_env      : dict,
                  use_spack_env  : bool,
+                 only_data      = False,
                  install        = False,
                  ):
         
-        self.handler_id = handler_id
-        self.spack_env  = spack_env
-        self.env_name   = env_name
-        self.use_spack_env = use_spack_env
+        self.initialized    = False
+        self.handler_id     = handler_id
+        self.spack_env      = spack_env
+        self.env_name       = env_name
+        self.__use_spack_env= use_spack_env
+        self.__only_data    = only_data
+        self.install        = install
         
-        self.compiler   = self.spack_env["compiler"]
-        self.target     = self.spack_env["target"]
-        self.language   = self.spack_env["language"]
-        self.packages   = self.spack_env["packages"]
+        self.compiler       = self.spack_env["compiler"]
+        self.target         = self.spack_env["target"]
+        self.language       = self.spack_env["language"]
+        self.packages       = self.spack_env["packages"]
+        self.package_locations = {}
         self.python_version = "python"
         
         self.additional = ""
@@ -53,21 +60,45 @@ class SpackManager:
         self.env_location.mkdir(parents=True, exist_ok=True)
         
         self.file_location = Path(f"{self.env_location.absolute()}/env-{self.env_name}.sh")
+        
+        
+        for name, info in self.packages.items():
+
+            versions = info["versions"] if type(info["versions"]) != str else [info["versions"]]
+            variants = [""]
+
+            if "variants" in info:
+                variants = info["variants"] if type(info["variants"]) != str else [info["variants"]]
+
+            combinations = itertools.product(*[versions, variants, [self.compiler]])
+                
+            for combination in combinations:
+                package = f"{name}@{combination[0]} {combination[1]} %{combination[2]}"
+                    
+                if name == "python":
+                    self.python_version = package
+                
+                if self.__only_data == False:
+                    p = subprocess.run(f"spack location -i {package}".split(), text=True, check=True, capture_output=True)
+                    self.package_locations[name] = (f"{package}", p.stdout.rstrip())
+                    
+            self.loadables[name] = deepcopy(combinations)
+        
+        
+        if Path(f"{self.env_location}/.venv").is_dir() == True:
+            self.initialized = True
+        
+
+    def initialize_env(self):
         with open(self.file_location, "w") as file:
             file.write(f"#!/bin/bash \n")
             
             first = True
             for index, (name, info) in enumerate(self.packages.items()):
 
-                versions    = info["versions"] if type(info["versions"]) != str else [info["versions"]]
-                variants = [""]
+                combinations = self.loadables[name]
 
-                if "variants" in info:
-                    variants= info["variants"] if type(info["variants"]) != str else [info["variants"]]
-
-                combinations = itertools.product(*[versions, variants, [self.compiler]])
-
-                if install == True:
+                if self.install == True:
 
                     fresh = ""
                     if "fresh" in info:
@@ -84,8 +115,7 @@ class SpackManager:
                 
                 if index == len(self.packages.items())-1:
                     file.write(f"spack -e {self.env_name} install\n")
-                
-                self.loadables[name] = deepcopy(combinations)
+                    
             
             file.write("spack env list\n")
             file.write(f"spack env activate {self.env_name}\n")
@@ -103,14 +133,17 @@ class SpackManager:
         
         print(bcolors.OKCYAN + f"Initialize environment {self.env_name}" + bcolors.ENDC)
         
-        if self.use_spack_env == True:
+        if self.__use_spack_env == True:
             p = subprocess.Popen(["bash", self.file_location], stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
             #for line in iter(lambda: p.stdout.readline(1), b""): # type: ignore
             #    sys.stdout.buffer.write(line)
             p.wait()
+            if p.returncode != 0:
+                raise RuntimeError(bcolors.FAIL + "Environment creation has failed" + bcolors.ENDC)
+            
         print(bcolors.OKGREEN + "Finish Initialize environment" + bcolors.ENDC)
-
-             
+        self.initialized = True
+          
         self.file_location.unlink()
    
    
@@ -119,8 +152,7 @@ class SpackManager:
 
         for name, combinations in self.loadables.items():
             loadable = self.__load_packages(loadable=loadable, combinations=deepcopy(combinations), package_name=name)
-         
-        #print(loadable)   
+          
         return loadable
         
         
@@ -135,14 +167,7 @@ class SpackManager:
         for combination in combinations:
             
             #print(f"spack -e {self.env_name} add {package_name}@{combination[0]} {combination[1]} %{combination[2]}")
-            package  = f"{package_name}@{combination[0]} {combination[1]} %{combination[2]}"
-            file.write(f"spack -e {self.env_name} add {package}\n")  
-            
-            if package_name == "python":
-                self.python_version = package
-            
-            p = subprocess.run(f"spack location -i {package}".split(), text=True, check=True, capture_output=True)
-            self.packages[package_name]["location"] = (f"{package}", p.stdout.rstrip())
+            file.write(f"spack -e {self.env_name} add {package_name}@{combination[0]} {combination[1]} %{combination[2]}\n")  
             
 
     def __load_packages(self, loadable: str, combinations, package_name: str):
@@ -155,17 +180,22 @@ class SpackManager:
             
            
     def delete(self):
-        with open(self.file_location, "w") as file:
-            file.write("#!/bin/bash  \n")
-            file.write(f"spack env activate {self.env_name} -p \n")   
-            file.write(f"spack remove --all \n")   
-            file.write(f"spack env deactivate\n")    
-            file.write(f"spack env remove {self.env_name} -y \n")
-                       
         
-        p = subprocess.run(["bash", self.file_location], capture_output=True, check=True)
-        #print(p.stderr)
-        print(p.stdout)
-        
-        self.file_location.unlink()
-        shutil.rmtree(path=self.env_location.absolute())
+        if self.initialized == True:
+            with open(self.file_location, "w") as file:
+                file.write("#!/bin/bash  \n")
+                file.write(f"spack env activate {self.env_name} -p \n")   
+                file.write(f"spack remove --all \n")   
+                file.write(f"spack env deactivate\n")    
+                file.write(f"spack env remove {self.env_name} -y \n")
+
+
+            p = subprocess.run(["bash", self.file_location], capture_output=True, check=True)
+            #print(p.stderr)
+            print(p.stdout)
+
+            self.file_location.unlink()
+            shutil.rmtree(path=self.env_location.absolute())
+            
+        else: 
+            print(bcolors.WARNING + f"{self.env_name} has not been initialized, nothing to delete" + bcolors.ENDC)
